Improve playback starting algorithm
We now wait until one of the queues is full. That should make sure that we have enough frames buffered at the beginning for a smooth playback and it should also make sure that we don't cause a deadlock if one of the queues gets filled up without having anything in the other queue.