Move audiostream code to FFmpeg
This reduces the direct dependency on the WAV and Ogg/Vorbis code. Since
that code is already present in FFmpeg, using that library improves our
code since we have to deal with fewer format specific code and probably
also fixes some issue with oddly encoded files.

The rest of the audio code still uses Wav and Ogg directly so that code
and libraries can't be removed yet.