Improve version detection
These changes are only required for cases where the system FFmpeg
library was updated or if there are outdated prebuilt libraries present.
In any case, the current minor version must be less than the compiled
version since that ensures that the ABI is still intact. Since a major
version bump means that the ABI is changes we require that the same
version is used.