Generialize graphics buffer binding
For another project I need to bind multiple buffers at the same
time but the current API does not allow that yet. I made that possible
by introducing a new class for storing the binding information of all
vertex buffer binding points.

I chose not to integrate this into the existing vertex_layout class since
that is only responsible for specifying the data layout and not the data
source.